requirements.txt
faiss_cpu==1.7.4
infomap==2.7.1
matplotlib==3.8.0
numpy==1.26.2
Pillow==10.0.1
Pillow==10.1.0
scikit_learn==1.3.1
scipy==1.11.4
seaborn==0.13.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
